Task 4 - Extra work/ Inventory system
Implemented an inventory system as an extra task. The player can pick up three types of potions: Health, Speed, and Immunity to Damage, by pressing E. They are then stored in the inventory for later use. The inventory can be opened with I ( I made Inventory canvas), and potions can be used anytime by pressing "Use Potion" buttons in the canvas. I also added a tooltip system that displays the remaining quantity of each potion when hovering over its card in the inventory.